requirements.txt
fastapi 
uvicorn 
requests

## Auto Instrumentation
opentelemetry-instrumentation
opentelemetry-distro
opentelemetry-exporter-otlp

## Manual Instrumentation - custom traces
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
structlog



